Rename before middleware, move middlewares to folder
