+++ src/metadata/types/before.metadata.ts
@@ -1,6 +1,6 @@
+import { BeforeMiddleware } from '../../middleware';
 import { FieldArgumentMetadata } from './field-argument.metadata';
-import { Middleware } from '../../middleware';
 
 export interface BeforeMetadata extends FieldArgumentMetadata {
-  middleware: Middleware;
+  middleware: BeforeMiddleware;
 }
