+++ src/middleware/after.middleware.ts
@@ -1,5 +1,3 @@
-export type Middleware = (context: any, args: { [key: string]: any }, next: (error?: Error, value?: any) => any) => Promise<any> | any;
-
 export type AfterMiddleware = (
   context: any,
   args: { [key: string]: any },
