+++ src/middleware/before.middleware.ts
@@ -1,0 +1,5 @@
+export type BeforeMiddleware = (
+  context: any,
+  args: { [key: string]: any },
+  next: (error?: Error, value?: any) => any,
+) => Promise<any> | any;
